perf(memory): optimize graph candidate generation and enforce LOC limits
