+++ crates/csm-memory/src/singularity.rs
@@ -491,7 +491,8 @@ impl<H: Hypervector + 'static> Singularity<H> {
                 }
             }
         }
-        incoming.sort_unstable_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
+        incoming
+            .sort_unstable_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
         incoming
     }
